+++ FRCore/FRCore/Security/SSHDetector.swift
@@ -11,6 +11,7 @@
 import Foundation
 
 /// SSHDetector is a JailbreakDetector class, and is used as one of default JailbreakDetector's detectors to determine whether the device is Jailbroken or not
+@available(*, deprecated, message: "Use SuspiciousFilesExistenceDetector and SuspiciousFilesAccessibleDetector instead")
 public class SSHDetector: JailbreakDetector {
     
     /// Initializes SSHDetector instance